+++ PandaSQL/Storage/BerkeleyDB/BDBSeqScanIterator.cpp
@@ -187,7 +187,7 @@ bool BDBSeqScanIterator::Update(const ValueList &inValueList)
 	std::string rowString;
 	TupleToString(mTupleDesc, inValueList, &rowString);
 	data.data = (void *)rowString.c_str();
-	data.size = rowString.length();
+	data.size = (u_int32_t)rowString.length();
 
 	ret = mpDBCursor->put(mpDBCursor, &key, &data, DB_CURRENT);
 
@@ -272,7 +272,7 @@ bool BDBSeqScanIterator::PutCursor_Private(const ValueList &inValueList, u_int32
 	std::string rowString;
 	TupleToString(mTupleDesc, inValueList, &rowString);
 	data.data = (void *)rowString.c_str();
-	data.size = rowString.length();
+	data.size = (u_int32_t)rowString.length();
 
 	ret = mpDBCursor->put(mpDBCursor, &key, &data, flags);
